# Conflicts:
#	registry-integration-tests/src/test/java/org/gbif/registry/ws/it/collections/resource/CollectionResourceIT.java
#	registry-persistence/src/main/resources/liquibase/master.xml
#	registry-service/src/main/java/org/gbif/registry/service/collections/descriptors/DefaultDescriptorService.java

+++ registry-service/src/main/java/org/gbif/registry/service/collections/descriptors/DefaultDescriptorService.java
@@ -236,14 +236,12 @@ public class DefaultDescriptorService implements DescriptorsService {
   @Override
   public void updateDescriptorGroup(
       @NotNull long descriptorGroupKey,
-      @NotNull byte[] descriptorGroupFile,
+      byte[] descriptorGroupFile,
       @NotNull ExportFormat format,
       @NotNull String title,
       Set<String> tags,
       String description
 ) {
-    Objects.requireNonNull(descriptorGroupFile);
-    Preconditions.checkArgument(descriptorGroupFile.length > 0);
     Preconditions.checkArgument(!Strings.isNullOrEmpty(title));
 
     // Validate tags against vocabulary server
@@ -266,13 +264,13 @@ public class DefaultDescriptorService implements DescriptorsService {
       descriptorGroup.setTags(tags);
     }
     descriptorsMapper.updateDescriptorGroup(descriptorGroup);
+    if(descriptorGroupFile != null ) {
+      // remove descriptors
+      descriptorsMapper.deleteDescriptors(descriptorGroup.getKey());
 
-    // remove descriptors
-    descriptorsMapper.deleteDescriptors(descriptorGroup.getKey());
-
-    // reimport the file
-    importDescriptorsFile(descriptorGroupFile, format, descriptorGroup.getKey());
-
+      // reimport the file
+      importDescriptorsFile(descriptorGroupFile, format, descriptorGroup.getKey());
+    }
     eventManager.post(
         SubEntityCollectionEvent.newInstance(
             descriptorGroup.getCollectionKey(),
